[Feature] Add a new module `vectorCgraConverter.js` to handle the conversion of architecture JSON to a VectorCGRA-compatible YAML format, supporting heterogeneous CGRA configurations.
